+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 This log documents the significant changes for each release.
 This project follows [Semantic Versioning](http://semver.org/).
 
+## [0.11.0] - 2019-05-21 (0.10.0 skipped)
+### Added
+- Questionnaire data extraction for Observation data is now supported via the
+  questionnaire-observationLinkPeriod extension.
+
 ## [0.9.1] - 2019-05-03
 ### Fixed
 Updated LForms to get several fixes: